refactor(plugin-server): only run App methods in a worker
At the moment we're running the whole pipeline in a worker. This is
great for CPU bound tasks, but not so useful for IO bound tasks.

The reason we still run the App methods in a worker is just to give us a
bit of isolation from badly behaving App code. This way we avoid using
the main thread event loop for this code, so worker threads can lock but
the main thread can still keep running. Piscina should hopefully reap
the problematic worker threads and respawn :fingercrossed:

Reducing the amount running in the worker thread helps e.g. with
debugging the plugin server, plays a little more nicely with e.g.
coverage reports (although it's possible to get coverage reports from
workers too), and makes it easier to reason about the code.